+++ libc/config/linux/riscv/headers.txt
@@ -31,6 +31,7 @@ set(TARGET_PUBLIC_HEADERS
     libc.include.sys_mman
     libc.include.sys_prctl
     libc.include.sys_random
+    libc.include.sys_queue
     libc.include.sys_resource
     libc.include.sys_select
     libc.include.sys_socket
