+++ libc/config/linux/x86_64/headers.txt
@@ -30,6 +30,7 @@ set(TARGET_PUBLIC_HEADERS
     libc.include.sys_ioctl
     libc.include.sys_mman
     libc.include.sys_prctl
+    libc.include.sys_queue
     libc.include.sys_random
     libc.include.sys_resource
     libc.include.sys_select
