+++ libc/include/CMakeLists.txt
@@ -351,7 +351,7 @@ add_header(
     sys/queue.h
   DEPENDS
     .llvm-libc-macros.null_macro
-    .llvm-libc-macros.offsetof_macro
+    .llvm-libc-macros.containerof_macro
 )
 
 add_gen_header(
