+++ libc/include/llvm-libc-macros/CMakeLists.txt
@@ -4,7 +4,7 @@ function(add_macro_header name)
     "MACRO_HEADER"
     "" # Optional arguments
     "HDR" # Single value arguments
-    "" # Multi-value arguments
+    "DEPENDS" # Multi-value arguments
     ${ARGN}
   )
   if(TARGET libc.include.llvm-libc-macros.${LIBC_TARGET_OS}.${name})
@@ -14,12 +14,15 @@ function(add_macro_header name)
         ${MACRO_HEADER_HDR}
       DEPENDS
         .${LIBC_TARGET_OS}.${name}
+        ${MACRO_HEADER_DEPENDS}
     )
   else()
     add_header(
       ${name}
       HDR
         ${MACRO_HEADER_HDR}
+      DEPENDS
+        ${MACRO_HEADER_DEPENDS}
     )
   endif()
 endfunction(add_macro_header)
@@ -77,6 +80,14 @@ add_macro_header(
 )
 
 add_macro_header(
+  containerof_macro
+  HDR
+    containerof-macro.h
+  DEPENDS
+    .offsetof_macro
+)
+
+add_macro_header(
   sched_macros
   HDR
     sched-macros.h
@@ -122,6 +133,12 @@ add_macro_header(
   sys_mman_macros
   HDR
     sys-mman-macros.h
+)
+
+add_macro_header(
+  sys_queue_macros
+  HDR
+    sys-queue-macros.h
 )
 
 add_macro_header(
