+++ libc/test/include/sys/queue_test.cpp
@@ -6,11 +6,10 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "llvm-libc-macros/sys-queue-macros.h"
 #include "src/__support/CPP/string.h"
 #include "src/__support/char_vector.h"
 #include "test/UnitTest/Test.h"
-
-#include <sys/queue.h>
 
 using LIBC_NAMESPACE::CharVector;
 using LIBC_NAMESPACE::cpp::string;
